refactor: extract ActionSection to separate component
